Transferred features from Toaster
Includes:
*Select a conversation from "Your Conversations", which will display in the Active Conversation box.

* When sending a message, switch to that conversation, and refresh it to show that new message

*GetConversation usage added (was previously defined in server but not used in client).

____ Unsuccessful Feature Transfers ___
* Auto-update conversation every 2.5s appears to bug out if the user has selected multiple conversations, it will cycle between them repeatedly. Disabled for now

Known Issues:
When two users are logged in at once the website cannot display them both at the same time.
ie, Upon page refresh, the whoever logged in first will have their list of conversations overwritten with the second user to log in.
Also When sending a message when two users are logged in, the username of the sender will always be the 2nd person to log in.